FinBot/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
ansi2html==1.8.0
appdirs==1.4.4
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
blis==0.7.9
bokeh==3.2.0
bs4==0.0.1
cachetools==5.3.1
catalogue==2.0.8
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
confection==0.0.4
contourpy==1.1.0
cymem==2.0.7
dash==2.11.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dataclasses==0.6
dataclasses-json==0.5.8
decorator==5.1.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl#sha256=0964370218b7e1672a30ac50d72cdc6b16f7c867496f1d60925691188f4d2510
et-xmlfile==1.1.0
faiss-cpu==1.7.4
Flask==2.2.5
frozendict==2.3.8
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
google-search-results==2.4.2
html5lib==1.1
idna==3.4
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
langchain==0.0.240
langchainplus-sdk==0.0.17
langcodes==3.3.0
langsmith==0.0.14
lxml==4.9.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
mdurl==0.1.2
multidict==6.0.4
multitasking==0.0.11
murmurhash==1.0.9
mypy-extensions==1.0.0
nest-asyncio==1.5.6
newsapi-python==0.2.7
numexpr==2.8.4
numpy==1.25.0
openai==0.27.8
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
packaging==23.1
pandas==2.0.2
pandas-ta==0.3.14b0
pathy==0.10.2
Pillow==9.5.0
plotly==5.15.0
preshed==3.0.8
promptlayer==0.1.90
protobuf==4.23.3
pyarrow==12.0.1
pydantic==1.10.9
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
PyPDF2==3.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2023.8.8
requests==2.31.0
retrying==1.3.4
rich==13.4.2
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
soupsieve==2.4.1
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==2.0.17
srsly==2.4.6
streamlit==1.24.0
streamlit-option-menu==0.3.6
tenacity==8.2.2
thinc==8.1.10
tiktoken==0.4.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
typer==0.7.0
typing-inspect==0.9.0
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.3
validators==0.20.0
wasabi==1.1.2
webencodings==0.5.1
Werkzeug==2.2.3
xyzservices==2023.5.0
yarl==1.9.2
yfinance==0.2.25
zipp==3.15.0
